mlops/requirements.txt
tensorflow
numpy
pandas
scikit-learn
joblib
seaborn
matplotlib
pymongo==4.0.0
pandas==1.3.3
python-dotenv==1.0.0
